refactor: 제보하기로 이동하는 float 이미지 공통 컴포넌트 처리 #12
