+++ src/features/safe-route/ui/SafeRouteMap.tsx
@@ -5,8 +5,7 @@ import { fitToCoords } from "../lib/fitBounds";
 import type { Loc } from "@shared/types/location";
 import { reverseGeocode } from "../lib/reverseGeocode";
 import markerIcon from "@shared/assets/icons/marker.png";
-import { useNavigate } from "react-router-dom";
-import styled from "styled-components";
+import FloatAction from "@shared/ui/FloatAction";
 
 type Props = {
   origin?: Loc;
@@ -279,30 +278,10 @@ export default function SafeRouteMap({
     }
   }
 
-  const navigate = useNavigate();
-  const onFloatClick = () => navigate("/citizenIntro");
-
   return (
     <>
       <div ref={containerRef} style={{ width: "100%", height: "100%" }} />
-      {!(origin && dest) && (
-        <FloatImg
-          src="/images/icons/mapFloating.png"
-          role="button"
-          onClick={onFloatClick}
-        />
-      )}
+      {!(origin && dest) && <FloatAction />}
     </>
   );
 }
-
-const FloatImg = styled.img`
-  width: 3.5rem;
-  height: 3.5rem;
-  position: absolute;
-  bottom: 80px;
-  right: 0;
-  cursor: pointer;
-  z-index: 40;
-  pointer-events: auto;
-`;
